% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{addLabel}
\alias{addLabel}
\alias{addLabels}
\title{Add labels to a story}
\usage{
addLabel(story, labels = character(0))

addLabels(story, labels = character(0))
}
\arguments{
\item{story}{A \code{story} object, a story id, or URL to a story}

\item{labels}{Character vector of names of labels to add to the story}
}
\value{
Invisibly, the \code{story} object with the labels added. If the story
already has all of the labels you're trying to add, no request will be made.
}
\description{
Add labels to a story
}
