% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{getStories}
\alias{getStories}
\title{Get stories}
\usage{
getStories(..., search = NULL, query = list())
}
\arguments{
\item{...}{"Filter" terms to refine the query. See \code{https://www.pivotaltracker.com/help/articles/advanced_search/}}

\item{search}{A search string}

\item{query}{List of query parameters. See \code{https://www.pivotaltracker.com/help/api/rest/v5#Stories}. Most are not valid when filter terms are used.}
}
\value{
A list of Stories
}
\description{
Get stories
}
