% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\title{Get a story's labels}
\usage{
labels(story)
}
\arguments{
\item{story}{A \code{story} object, a story id, or URL to a story}
}
\value{
A character vector of the names of the labels attached to the story
}
\description{
Get a story's labels
}
